Successful implementation of the Matlab to the PSM code. Still differences between my results and Annette's.
MODIFIED FILES:
    BEHRDaily_Map.py - primary point of access for the Matlab interface.
Changes:
    -- New function imatlab_gridding serves as the gateway function that Matlab
should call to access the gridding functionality.
    -- Basemap and pyplot no longer imported, plotting disabled. Basemap is not
included in system Python, just in IPython.
    -- Necessary variable returns added
    -- Some runtime changes for the testing. Grid is currently set to
"northamerica_behrtest" which matches the grid Annette used. Restore to
"northamerica_behr" to match the normal BEHR grid.

    average_daybyday.py - mostly run time changes, new file paths, etc.

    omi/omi/data/grids.json - added "northamerica_behrtest" grid.

NEW FILES:
    psm_wrapper.m - The Matlab side of the interface. Handles conversion of a
Data structure to the Python list of dicts required to be passed to the PSM
algorithm, and the reverse (converting the output grid to a Matlab structure).
Does some checking on the Data structure before converting it; ensures that the
required fields are present and that all fill values are converted to NaNs.