PSM algorithm works with BEHR data and the new weights
MODIFIED FILES:
    PSM_Main.py:
        ++ catch QHullErrors and skip that swath
        ++ debugging messages controlled by __verbosity__
        ++ filtering for invalid AMFs (< 10^-5)
        ++ properly check if vals returned from grid_orbits not None before
adding to cumulative grid values
        ++ multi-day averages implemented as function
        ++ datasets squeezed to ensure 1D datasets are actually 1D, which fixes
the issues with clip_orbit
        ++ grid_orbit now correctly returns two "None" objects when aborting an
orbit, this fixes the "None type not iterable" error

    omi/omi/__init__.py:
        ++ clip_orbit now gets the along and across track lengths from the
longitude array, assuming that its dimensions are along x across or other x
along x across.
        ++ remove_out_of_domain_data now takes those along and across lengths
and uses them to figure out how to clip each dataset, this removes some of the
guesswork that was there previously.

    omi/omi/he5.py:
        ++ read_datasets now checks the fillvalue property directly present on each field (which is what BEHR sets instead of the attribute _FillValue)
        ++ read_datasets also has a keyword argument which causes it to
transpose arrays when read in, which gets BEHR data back to the SP orientation