Hotfix of error when along track dimension == 4
When along track dimension == 4, the previous check in remove_out_of_domain_data
that a 3D array is ordered [corner, along, across] or [along, across, corner] failed
because it would get the index of the corner dimension instead of the along dimension.
Now I've fixed it so that it checks that the dimension before the across track one
is the right size for the along track one.