Added cvm_generic_wrapper.m and the supporting functions in PSM_Main.py.
psm_wrapper requires that a BEHR Data structure be passed into it, which
is great for the main BEHR algorithm and not so great if you want to grid
some normal quantity. The new cvm_generic_wrapper allows you to just grid
any normal array of data without needed to be in a BEHR data structure.